Removed function & dropdowns, credit changes & header location changes
I removed the hidden function for divs so it shows everything in the account portal. Removed drop downs as they have no purpose anymore. Made it so it now adds the amount input on top of the current account credit instead of having to set a specific number. Also changed some of the redirects.